Clean up after establishing variant style
